fix(mobile-sync): stop publishing an active group the phone was never sent
When every tab in the active group is unpublishable — a combined-diff tab,
a workspace-document browser tab — the group is held back from `tabGroups`,
but the snapshot still published `activeGroupId` naming it. The client was
told the active group is one it never received.

Resolve the published pointer against the published group set: keep it when
the group survived, else the first published group, else null. Both fallbacks
are values three of the three read sites already handle today.

The published invariants covered group->tab dangling references and the
snapshot's own `activeTabId`, but nothing checked `activeGroupId` against the
published groups. The extended invariant now checks every published group
reference — `activeGroupId` and the `tabGroupLayout` leaves — and moves to its
own file so the claim has a name.

STA-5724
